Move to ExecutionDevice trait
