+++ src-tauri/Cargo.toml
@@ -26,6 +26,7 @@ rand = "0.8.5"
 rand_chacha = "0.3.1"
 sha2 = "0.10.6"
 uuid = "1.3.0"
+async-trait = "0.1.68"
 
 titan = { git = "https://github.com/1whatleytay/titan.git", branch = "main" }
